// DXC_Frontend/src/api/models/updateSurveyCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface UpdateSurveyCommand {
  id?: number
  /** @nullable */
  tenKhaoSat: string | null
  thoiGian?: string
  /** @nullable */
  displayWebsite?: string | null
  /** @nullable */
  header?: string | null
  /** @nullable */
  footer?: string | null
  /** @nullable */
  veViec?: string | null
  /** @nullable */
  isActive?: boolean | null
}