// DXC_Frontend/src/api/models/surveyDetailDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { QuestionDto } from './questionDto'

export interface SurveyDetailDto {
  id?: number
  /** @nullable */
  tenKhaoSat?: string | null
  thoiGian?: string
  /** @nullable */
  displayWebsite?: string | null
  /** @nullable */
  header?: string | null
  /** @nullable */
  footer?: string | null
  /** @nullable */
  veViec?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  questions?: QuestionDto[] | null
}